+++ CHANGELOG.md
@@ -14,6 +14,7 @@
   - Improved dataset creation and access by organization editors and viewers
   - Improved organization reports to show breakdowns of submitted and approved
     data requests, as well as overall downloads
+  - Data requests can now be exported by organization editors and viewers
 - **Variable Changes**
   - Remove extra space in variables index placeholder text
   - Move file download link to display below inline PDF on variable form page